Move `PmfSampling` to the `Sampling` object
